+++ exonum/src/node/mod.rs
@@ -43,7 +43,7 @@ use messages::{Connect, Message, RawMessage};
 use events::{NetworkRequest, TimeoutRequest, NetworkEvent, InternalRequest, InternalEvent,
              SyncSender, HandlerPart, NetworkConfiguration, NetworkPart, InternalPart};
 use events::error::{into_other, other_error, LogError, log_error};
-use helpers::{Height, Milliseconds, Round, ValidatorId};
+use helpers::{Height, Milliseconds, Round, ValidatorId, user_agent};
 use storage::Database;
 
 pub use self::state::{RequestData, State, TxPool, ValidatorState};
@@ -404,6 +404,7 @@ impl NodeHandler {
             &config.listener.consensus_public_key,
             external_address,
             system_state.current_time(),
+            &user_agent::get(),
             &config.listener.consensus_secret_key,
         );
 
@@ -486,7 +487,11 @@ impl NodeHandler {
             info!("Trying to connect with peer {}", address);
         }
 
-        let round = Round::first();
+        let snapshot = self.blockchain.snapshot();
+        let schema = Schema::new(&snapshot);
+
+        // Recover previous saved round if any
+        let round = schema.consensus_round();
         self.state.jump_round(round);
         info!("Jump to round {}", round);
 
@@ -494,6 +499,13 @@ impl NodeHandler {
         self.add_status_timeout();
         self.add_peer_exchange_timeout();
         self.add_update_api_state_timeout();
+
+        // Recover cached consensus messages if any. We do this after main initialization and before
+        // the start of event processing.
+        let messages = schema.consensus_messages_cache();
+        for msg in messages.iter() {
+            self.handle_message(msg);
+        }
     }
 
     /// Sends the given message to a peer by its id.